+++ dali/pipeline/executor/executor_impl.h
@@ -31,13 +31,13 @@
 #include "dali/core/nvtx.h"
 #include "dali/pipeline/executor/executor.h"
 #include "dali/pipeline/data/backend.h"
+#include "dali/pipeline/executor/op_graph_storage.h"
+#include "dali/pipeline/executor/op_graph_verifier.h"
 #include "dali/pipeline/executor/queue_metadata.h"
 #include "dali/pipeline/executor/queue_policy.h"
 #include "dali/pipeline/executor/workspace_policy.h"
 #include "dali/pipeline/executor/iteration_data.h"
 #include "dali/pipeline/graph/op_graph.h"
-#include "dali/pipeline/graph/op_graph_storage.h"
-#include "dali/pipeline/graph/op_graph_verifier.h"
 #include "dali/pipeline/operator/batch_size_provider.h"
 #include "dali/pipeline/operator/builtin/conditional/split_merge.h"
 #include "dali/pipeline/operator/checkpointing/checkpoint.h"
